cambios charlados

+++ TpIngSoftII/Interfaces/Services/IEmpleadoService.cs
@@ -11,7 +11,7 @@ namespace TpIngSoftII.Interfaces.Services
     public interface IEmpleadoService : IEntityAppServiceBase<Empleado, EmpleadoDto>
     {
         int ValidarCredenciales(LoginRequest login);
-        decimal Antiguedad(int empleadoID);
+        int Antiguedad(int empleadoID);
         EmpleadoDto DameMisDatos();
         int GetEmpleadoUsuarioPassword(string nombreUsuario, string passwordUsuario);
         IEnumerable<PerfilDto> GetPerfilesDeEmpleado(int empleadoID);
